Links API JWT authorization.

+++ Controllers/API/LinksApiController.cs
@@ -4,8 +4,11 @@ using Linkeeper.DTOs;
 using Linkeeper.Models;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using System.Collections.Generic;
+using System.Threading.Tasks;
 
 namespace Linkeeper.Controllers
 {
@@ -16,56 +19,81 @@ namespace Linkeeper.Controllers
 	{
 		private readonly ILinkeeperRepo _repository;
 		private readonly IMapper _mapper;
+		private readonly UserManager<IdentityUser> _userManager;
 
-		public LinksApiController(ILinkeeperRepo repository, IMapper mapper)
+		public LinksApiController(ILinkeeperRepo repository, IMapper mapper, UserManager<IdentityUser> userManager)
 		{
 			_repository = repository;
 			_mapper = mapper;
+			_userManager = userManager;
 		}
 
 		[HttpGet]
-		public ActionResult<IEnumerable<Link>> GetAllLinks()
+		public async Task<ActionResult<IEnumerable<Link>>> GetAllUserLinksAsync()
 		{
-			var links = _repository.GetAllLinks();
+			string userId = "";
+			foreach (var claim in User.Claims) if (claim.Type == "Id") userId = claim.Value;
+			IdentityUser user = await _userManager.FindByIdAsync(userId);
+
+			var links = _repository.GetAllUserLinks(user);
+			if(links == null) return Ok();
 			return Ok(_mapper.Map<IEnumerable<LinkReadDTO>>(links));
 		}
 
 		[HttpGet("{id}", Name = "GetLinkById")]
-		public ActionResult<Link> GetLinkById(int id)
+		public async Task<ActionResult<Link>> GetLinkByIdAsync(int id)
 		{
+			string userId = "";
+			foreach (var claim in User.Claims) if (claim.Type == "Id") userId = claim.Value;
+			IdentityUser user = await _userManager.FindByIdAsync(userId);
+
 			Link link = _repository.GetLinkById(id);
+
 			if(link != null)
 			{
+				if (link.User == null) return StatusCode(StatusCodes.Status403Forbidden);
 				return Ok(_mapper.Map<LinkReadDTO>(link));
 			}
 			return NotFound();
 		}
 
 		[HttpPost]
-		public ActionResult<LinkReadDTO> AddLink(LinkAddDTO linkAddDTO)
+		public async Task<ActionResult<LinkReadDTO>> AddLinkAsync(LinkAddDTO linkAddDTO)
 		{
+			string userId = "";
+			foreach (var claim in User.Claims) if (claim.Type == "Id") userId = claim.Value;
+			IdentityUser user = await _userManager.FindByIdAsync(userId);
+
 			Link link = _mapper.Map<Link>(linkAddDTO);
+			link.User = user;
 			_repository.AddLink(link);
 			bool added = _repository.SaveChanges();
 
 			if (added)
 			{
 				LinkReadDTO linkReadDTO = _mapper.Map<LinkReadDTO>(link);
-				return CreatedAtRoute(nameof(GetLinkById), new { id = linkReadDTO.Id }, linkReadDTO);
+				return CreatedAtRoute(nameof(GetLinkByIdAsync), new { id = linkReadDTO.Id }, linkReadDTO);
 			}
 
 			return BadRequest();
 		}
 
 		[HttpPut("{id}")]
-		public ActionResult UpdateLink(int id, LinkUpdateDTO linkUpdateDTO)
+		public async Task<ActionResult> UpdateLinkAsync(int id, LinkUpdateDTO linkUpdateDTO)
 		{
+			string userId = "";
+			foreach (var claim in User.Claims) if (claim.Type == "Id") userId = claim.Value;
+			IdentityUser user = await _userManager.FindByIdAsync(userId);
+
 			Link link = _repository.GetLinkById(id);
 
 			if (link == null) 
 				return NotFound();
 			
 			_mapper.Map(linkUpdateDTO, link);
+
+			if (link.User == null) return StatusCode(StatusCodes.Status403Forbidden);
+
 			_repository.UpdateLink(link);
 			bool added = _repository.SaveChanges();
 
@@ -76,12 +104,18 @@ namespace Linkeeper.Controllers
 		}
 
 		[HttpDelete("{id}")]
-		public ActionResult DeleteLink(int id)
+		public async Task<ActionResult> DeleteLinkAsync(int id)
 		{
+			string userId = "";
+			foreach (var claim in User.Claims) if (claim.Type == "Id") userId = claim.Value;
+			IdentityUser user = await _userManager.FindByIdAsync(userId);
+
 			Link link = _repository.GetLinkById(id);
 
 			if (link == null) 
 				return NotFound();
+
+			if (link.User == null) return StatusCode(StatusCodes.Status403Forbidden);
 
 			_repository.DeleteLink(link);
 			_repository.SaveChanges();
